+++ src/core/component/field/index.ts
@@ -11,7 +11,6 @@
  * @packageDocumentation
  */
 
-import { NULL } from 'core/component/const';
 import { ComponentInterface, ComponentField, ComponentSystemField } from 'core/component/interface';
 
 // Queue for fields to initialize
@@ -21,7 +20,7 @@ export const fieldQueue = new Set();
  * Initializes the specified fields to a component instance.
  * The function returns an object with initialized fields.
  *
- * This method has some "copy-paste" chunks, but it's done for better performance, because it's a very hot function.
+ * This method has some "copy-paste" chunks, but it's done for better performance, because it's a very "hot" function.
  * Mind that the initialization of fields is a synchronous operation.
  *
  * @param fields - component fields or system fields
@@ -34,63 +33,65 @@ export function initFields(
 	component: ComponentInterface,
 	store: Dictionary = {}
 ): Dictionary {
-	const
-		{unsafe} = component,
-		{meta: {params, instance}} = unsafe;
+	const {
+		unsafe,
+		unsafe: {meta: {params, instance}},
+		isFlyweight
+	} = component;
 
 	const
 		// True if a component is functional or a flyweight
-		isFlyweight = params.functional === true || unsafe.isFlyweight;
+		isNotRegular = params.functional === true || isFlyweight;
 
 	const
-		// Map of fields that we should skip, i.e. not to initialize.
-		// For instance, some properties isn't initialized if a component is functional.
-		fieldsToSkip = Object.createDict(),
+		// Map of fields that we should skip, i.e., not to initialize.
+		// For instance, some properties don't initialize if a component is a functional.
+		canSkip = Object.createDict(),
 
-		// List of atomics to initialize
-		atomList = <string[]>[],
+		// List of atoms to initialize
+		atomList = <Array<Nullable<string>>>[],
 
-		// List of non-atomics fields to initialize
-		fieldList = <string[]>[];
+		// List of non-atoms to initialize
+		nonAtomList = <Array<Nullable<string>>>[];
 
-	// At first we should initialize all atomic fields, but some atomics wait another atomics.
-	// That's why we sort list of fields and organize simple synchronous queue.
-	// All atomics that waits another atomics are added to atomList.
-	// All non-atomics fields are added to fieldList.
+	// At first, we should initialize all atoms, but some atoms wait for other atoms.
+	// That's why we sort the source list of fields and organize a simple synchronous queue.
+	// All atoms that waits for other atoms are added to `atomList`.
+	// All non-atoms are added to `nonAtomList`.
 	for (let keys = Object.keys(fields).sort(), i = 0; i < keys.length; i++) {
 		const
 			key = keys[i],
-			el = <ComponentSystemField>fields[key];
+			el = <Nullable<ComponentSystemField>>fields[key];
 
-		// Don't initialize a property for a functional component
-		// unless explicitly required (functional == false)
-		if (isFlyweight && el.functional === false) {
-			fieldsToSkip[key] = true;
+		const needInit =
+			el != null &&
+			(!isNotRegular || el.functional !== false) &&
+			(el.init != null || el.default !== undefined || instance[key] !== undefined) &&
+			(store[key] === undefined || isFlyweight && el.replace === true);
+
+		if (el == null || !needInit) {
+			canSkip[key] = true;
 			continue;
 		}
 
-		// If a field is atomic, or if the field doesn't have an initializer:
-		// if the field doesn't have the initializer it guarantees doesn't have any dependencies
-		if (el.atom || !el.init && (el.default !== undefined || key in instance)) {
+		if (el.atom) {
 			// If true, then the field doesn't have any dependencies and can be initialized right now
 			let canInit = true;
 
-			// Set of dependencies to wait
-			const {after} = el;
+			const
+				{after} = el;
 
 			if (after && after.size > 0) {
 				for (let o = after.values(), val = o.next(); !val.done; val = o.next()) {
 					const
-						waitFieldKey = val.value;
+						waitKey = val.value;
 
-					if (fieldsToSkip[waitFieldKey] === true) {
+					if (canSkip[waitKey] === true) {
 						continue;
 					}
 
-					// Check the dependency is not already initialized.
-					// NULL is a special value to prevent undefined behaviour
-					// when a property already initialized with an undefined value.
-					if (!(waitFieldKey in store) || store[waitFieldKey] === NULL) {
+					// Check the dependency is not already initialized
+					if (!(waitKey in store)) {
 						atomList.push(key);
 						canInit = false;
 						break;
@@ -99,22 +100,18 @@ export function initFields(
 			}
 
 			if (canInit) {
-				if (store[key] === NULL) {
-					store[key] = undefined;
-				}
-
 				Object.set(unsafe, '$activeField', key);
 
 				let
 					val;
 
-				if (el.init) {
+				if (el.init != null) {
 					val = el.init(unsafe, store);
 				}
 
 				if (val === undefined) {
 					if (store[key] === undefined) {
-						// We need to clone default value from a constructor
+						// We need to clone the default value from a constructor
 						// to prevent linking to the same object for a non-primitive value
 						val = el.default !== undefined ? el.default : Object.fastClone(instance[key]);
 						store[key] = val;
@@ -128,59 +125,46 @@ export function initFields(
 			}
 
 		} else {
-			fieldList.push(key);
+			nonAtomList.push(key);
 		}
 	}
 
-	// Initialize all atomics that have some dependencies
+	// Initialize all atoms that have some dependencies
 	while (atomList.length > 0) {
 		for (let i = 0; i < atomList.length; i++) {
 			const
-				key = atomList[i];
+				key = nonAtomList[i];
 
-			if (!Object.isTruly(key)) {
+			if (key == null || key in store) {
 				continue;
 			}
 
-			const
-				isNull = store[key] === NULL;
-
-			if (key in store && !isNull) {
-				continue;
-			}
-
-			const
-				el = <Nullable<ComponentSystemField>>fields[key];
-
-			if (!el) {
-				continue;
-			}
-
-			// If true, then all dependencies of the fields is already initialized, and we can initialize this field
+			// If true, then the field doesn't have any dependencies and can be initialized right now
 			let canInit = true;
 
-			// Set of dependencies to wait
-			const {after} = el;
+			const
+				el = <ComponentSystemField>fields[key],
+				{after} = el;
 
 			if (after && after.size > 0) {
 				for (let o = after.values(), val = o.next(); !val.done; val = o.next()) {
 					const
-						waitFieldKey = val.value,
-						waitField = fields[waitFieldKey];
+						waitKey = val.value,
+						waitFor = fields[waitKey];
 
-					if (fieldsToSkip[waitFieldKey] === true) {
+					if (canSkip[waitKey] === true) {
 						continue;
 					}
 
-					if (!waitField) {
-						throw new ReferenceError(`The field "${waitFieldKey}" is not defined`);
+					if (!waitFor) {
+						throw new ReferenceError(`The field "${waitKey}" is not defined`);
 					}
 
-					if (!waitField.atom) {
-						throw new Error(`The atom field "${key}" can't wait the non atom field "${waitFieldKey}"`);
+					if (!waitFor.atom) {
+						throw new Error(`The atom field "${key}" can't wait the non atom field "${waitKey}"`);
 					}
 
-					if (!(waitFieldKey in store) || store[waitFieldKey] === NULL) {
+					if (!(waitKey in store)) {
 						fieldQueue.add(key);
 						canInit = false;
 						break;
@@ -188,27 +172,25 @@ export function initFields(
 				}
 
 				if (canInit) {
-					atomList[i] = '';
+					atomList[i] = null;
 				}
 			}
 
 			if (canInit) {
-				if (isNull) {
-					store[key] = undefined;
-				}
-
 				Object.set(unsafe, '$activeField', key);
 				fieldQueue.delete(key);
 
 				let
 					val;
 
-				if (el.init) {
+				if (el.init != null) {
 					val = el.init(unsafe, store);
 				}
 
 				if (val === undefined) {
 					if (store[key] === undefined) {
+						// We need to clone the default value from a constructor
+						// to prevent linking to the same object for a non-primitive value
 						val = el.default !== undefined ? el.default : Object.fastClone(instance[key]);
 						store[key] = val;
 					}
@@ -221,57 +203,44 @@ export function initFields(
 			}
 		}
 
-		// All atomics are initialized
+		// All atoms are initialized
 		if (fieldQueue.size === 0) {
 			break;
 		}
 	}
 
 	// Initialize all non-atomics
-	while (fieldList.length > 0) {
-		for (let i = 0; i < fieldList.length; i++) {
+	while (nonAtomList.length > 0) {
+		for (let i = 0; i < nonAtomList.length; i++) {
 			const
-				key = fieldList[i];
+				key = nonAtomList[i];
 
-			if (!Object.isTruly(key)) {
+			if (key == null || key in store) {
 				continue;
 			}
 
 			const
-				isNull = store[key] === NULL;
+				el = <ComponentSystemField>fields[key],
+				{after} = el;
 
-			if (key in store && !isNull) {
-				continue;
-			}
-
-			const
-				el = <Nullable<ComponentSystemField>>fields[key];
-
-			if (!el) {
-				continue;
-			}
-
-			// If true, then all dependencies of the fields is already initialized, and we can initialize this field
+			// If true, then the field doesn't have any dependencies and can be initialized right now
 			let canInit = true;
-
-			// Set of dependencies to wait
-			const {after} = el;
 
 			if (after && after.size > 0) {
 				for (let o = after.values(), val = o.next(); !val.done; val = o.next()) {
 					const
-						waitFieldKey = val.value,
-						waitField = fields[waitFieldKey];
+						waitKey = val.value,
+						waitFor = fields[waitKey];
 
-					if (fieldsToSkip[waitFieldKey] === true) {
+					if (canSkip[waitKey] === true) {
 						continue;
 					}
 
-					if (!waitField) {
-						throw new ReferenceError(`The field "${waitFieldKey}" is not defined`);
+					if (!waitFor) {
+						throw new ReferenceError(`The field "${waitKey}" is not defined`);
 					}
 
-					if (!(waitFieldKey in store) || store[waitFieldKey] === NULL) {
+					if (!(waitKey in store)) {
 						fieldQueue.add(key);
 						canInit = false;
 						break;
@@ -279,27 +248,25 @@ export function initFields(
 				}
 
 				if (canInit) {
-					fieldList[i] = '';
+					nonAtomList[i] = null;
 				}
 			}
 
 			if (canInit) {
-				if (isNull) {
-					store[key] = undefined;
-				}
-
 				Object.set(unsafe, '$activeField', key);
 				fieldQueue.delete(key);
 
 				let
 					val;
 
-				if (el.init) {
+				if (el.init != null) {
 					val = el.init(unsafe, store);
 				}
 
 				if (val === undefined) {
 					if (store[key] === undefined) {
+						// We need to clone the default value from a constructor
+						// to prevent linking to the same object for a non-primitive value
 						val = el.default !== undefined ? el.default : Object.fastClone(instance[key]);
 						store[key] = val;
 					}