+++ src/core/component/field/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.??? (2021-??-??)
+
+#### :bug: Bug Fix
+
+* Fixed creation of fields of flyweight components
+
 ## v3.0.0-rc.46 (2020-07-31)
 
 #### :house: Internal